requirements.txt
django>=4.2
psycopg2-binary>=2.9
pandas>=2.0
openpyxl>=3.1
djangorestframework>=3.14
celery>=5.3
redis>=5.0
drf-yasg>=1.21
drf-spectacular>=0.27
django-filter>=24.1
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
channels>=4.0.0
daphne>=4.0.0

# Testing dependencies
factory-boy>=3.3.0
coverage>=7.0.0
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0  # Para tests paralelos
freezegun>=1.2.0     # Para mockear tiempo
responses>=0.23.0    # Para mockear HTTP requests
model-bakery>=1.12.0 # Alternativa a factory-boy
django-test-plus>=2.2.0  # Utilidades adicionales para testing